feat: move question participation component to a separate file
