+++ apps/client/components/organisms/ExamDetails/ExamParticipation/ExamParticipationForm.tsx
@@ -1,83 +1,12 @@
-import type { Answer, Question } from '@dinan/types';
+import { useState } from 'react';
 import { Button } from 'components/atoms/Button';
-import { Checkbox } from 'components/atoms/Checkbox';
-import React, { useCallback, useState } from 'react';
-import { ParticipantAnswer } from '@dinan/types/ExamParticipation';
-type toggleAnswerFn = (
-  questionId: string,
-  answerId: string,
-  isChecked: boolean
-) => void;
+import { QuestionParticipationForm } from './QuestionParticipationForm';
+import type { Question } from '@dinan/types';
+import type { ParticipantAnswer } from '@dinan/types/ExamParticipation';
 
 type ExamParticipationFormProps = {
   questions: Question[];
   examAddress: string;
-};
-
-type QuestionParticipationFormProps = {
-  question: Question;
-  questionIndex: number;
-  toggleAnswer: toggleAnswerFn;
-};
-
-type AnswerParticipationFormProps = {
-  answer: Answer;
-  answerIndex: number;
-  questionId: string;
-  toggleAnswer: toggleAnswerFn;
-};
-
-const AnswerParticipationForm = ({
-  answer,
-  answerIndex,
-  questionId,
-  toggleAnswer,
-}: AnswerParticipationFormProps) => {
-  const handleToggleAnswer = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const isAnswerChecked = event.target.checked;
-    toggleAnswer(questionId, answer.id, isAnswerChecked);
-  };
-
-  return (
-    <li className="flex gap-2 mb-2">
-      <div className="flex items-center gap-2 self-start">
-        <p className="w-2 text-primary font-bold">{answerIndex + 1}. </p>
-        <Checkbox
-          className="w-max mt-0 gap-0 mb-0"
-          onChange={handleToggleAnswer}
-        />
-      </div>
-      <p>{answer.text}</p>
-      <p>{answer.id}</p>
-    </li>
-  );
-};
-
-const QuestionParticipationForm = ({
-  question,
-  questionIndex,
-  toggleAnswer,
-}: QuestionParticipationFormProps) => {
-  return (
-    <li className="w-4/5 my-2 border border-secondary rounded-md flex gap-4 flex-col p-4">
-      <p className="text-justify">
-        <span className="text-primary font-bold">{questionIndex + 1}. </span>
-        <span>{question.text}</span>
-      </p>
-      <p>Answers: </p>
-      <ul>
-        {question.answers.map((answer, index) => (
-          <AnswerParticipationForm
-            key={answer.id}
-            answer={answer}
-            answerIndex={index}
-            toggleAnswer={toggleAnswer}
-            questionId={question.id}
-          />
-        ))}
-      </ul>
-    </li>
-  );
 };
 
 const getInitialParticipantAnswers = (
@@ -85,6 +14,8 @@ const getInitialParticipantAnswers = (
 ): ParticipantAnswer[] => {
   return questions.map((q) => ({ questionId: q.id, answerIds: [] }));
 };
+
+const useSubmitExam = (examAddress: string) => {};
 
 export const ExamParticipationForm = ({
   questions,
@@ -100,7 +31,6 @@ export const ExamParticipationForm = ({
     isChecked: boolean
   ) => {
     // you need to check whether user mark the answer for the first time or change their mind and it's need to be removed
-
     if (isChecked) {
       const newParticipantAnswers = participantAnswers.map((pa) => {
         if (pa.questionId === questionId)
